Added CLI option for skipping reduction size check. Added document object as result of parsing .csr file. Added DFA graph class and basic operations.

+++ src/includes/util/utils.hpp
@@ -6,7 +6,7 @@
 
 #define SIZEOF_ARRAY(a)				(sizeof(a)/sizeof(*a))
 
-#define VALID_INDEX(i, length)		(0 <= (int)(i) && (int)(i) < (int)(length))
+#define VALID_INDEX(i, length)		(0 <= (i) && (i) < (length))
 
 #define VALID_ARRAY_INDEX(i, a)		VALID_INDEX(i, SIZEOF_ARRAY(a))
